requirements.txt
yfinance
confluent-kafka
psycopg2-binary 
python-dotenv 